+++ react/components/comparisonList/ComparisonGrid.tsx
@@ -2,14 +2,28 @@ import React, { useMemo } from 'react'
 import { pathOr, findLast, propEq } from 'ramda'
 import ComparisonGridRow from './ComparisonGridRow'
 import ComparisonProductContext from './ComparisonProductContext'
+import {
+  getProductFields,
+  getSkuFields,
+  getProductSpecificationFields,
+  getSkuSpecificationFields,
+} from '../utils/fieldUtils'
 
 interface Props {
-  maxItemCount: number
   columnStyles: Styles
-  fields: ComparisonField[]
+  productFieldsToHide?: string
+  skuFieldsToHide?: string
+  productSpecificationsToHide?: string
+  skuSpecificationsToHide?: string
 }
 
-const ComparisonGrid = ({ columnStyles, fields }: Props) => {
+const ComparisonGrid = ({
+  columnStyles,
+  productFieldsToHide,
+  skuFieldsToHide,
+  productSpecificationsToHide,
+  skuSpecificationsToHide,
+}: Props) => {
   const { useComparisonProductState } = ComparisonProductContext
 
   const productData = useComparisonProductState()
@@ -26,11 +40,16 @@ const ComparisonGrid = ({ columnStyles, fields }: Props) => {
       )
     })
 
-    const specificationsList: string[] = allProductSpecificationsList.reduce(
+    let specificationsList: string[] = allProductSpecificationsList.reduce(
       (accumulator: string[], currentValue: string[]) => {
         return [...new Set([...accumulator, ...currentValue])]
       },
       [] as string[]
+    )
+
+    specificationsList = getProductSpecificationFields(
+      specificationsList,
+      productSpecificationsToHide
     )
 
     const productSpecificationFields = specificationsList.map(
@@ -43,18 +62,11 @@ const ComparisonGrid = ({ columnStyles, fields }: Props) => {
     )
 
     return productSpecificationFields
-  }, [products])
+  }, [productSpecificationsToHide, products])
 
   const skuSpecificationFields: ComparisonField[] = useMemo(() => {
     const allSkuSpecificationsList: string[][] = products.map(product => {
-      // const skus = pathOr([], ['items'], product)
-
-      // const skuSpecificationsList = skus.map(sku =>
-      //   pathOr([], ['skuSpecifications'], sku)
-      // )
-
       const skuSpecificationsList = pathOr([], ['skuSpecifications'], product)
-
       const skuSpecificationNamesList: string[] = skuSpecificationsList.reduce(
         (accumulator: string[], currentValue) => {
           return [
@@ -70,11 +82,16 @@ const ComparisonGrid = ({ columnStyles, fields }: Props) => {
       return skuSpecificationNamesList.filter(name => name !== '')
     })
 
-    const specificationsList: string[] = allSkuSpecificationsList.reduce(
+    let specificationsList: string[] = allSkuSpecificationsList.reduce(
       (accumulator: string[], currentValue: string[]) => {
         return [...new Set([...accumulator, ...currentValue])]
       },
       [] as string[]
+    )
+
+    specificationsList = getSkuSpecificationFields(
+      specificationsList,
+      skuSpecificationsToHide
     )
 
     const skuSpecificationFieldsList = specificationsList.map(
@@ -87,39 +104,15 @@ const ComparisonGrid = ({ columnStyles, fields }: Props) => {
     )
 
     return skuSpecificationFieldsList
-  }, [products])
+  }, [products, skuSpecificationsToHide])
 
-  fields =
-    fields && fields.length > 0
-      ? fields
-      : [
-          {
-            type: 'ProductField',
-            name: 'productName',
-            displayValue: 'Product Name',
-            showOnSite: true,
-          },
-          {
-            type: 'ProductField',
-            name: 'brand',
-            displayValue: 'Brand',
-            showOnSite: true,
-          },
-          {
-            type: 'ProductField',
-            name: 'description',
-            displayValue: 'Product Description',
-            showOnSite: true,
-          },
-          {
-            type: 'ProductField',
-            name: 'productReference',
-            displayValue: 'Product Reference',
-            showOnSite: true,
-          },
-          ...productSpecificationFields,
-          ...skuSpecificationFields,
-        ]
+  const fields: ComparisonField[] = [
+    ...getProductFields(productFieldsToHide),
+    ...getSkuFields(skuFieldsToHide),
+
+    ...productSpecificationFields,
+    ...skuSpecificationFields,
+  ]
 
   return fields.map((field: ComparisonField) => {
     return (
@@ -130,6 +123,53 @@ const ComparisonGrid = ({ columnStyles, fields }: Props) => {
       />
     )
   })
+}
+
+ComparisonGrid.schema = {
+  title: 'admin/editor.comparison-grid-row.title',
+  description: 'admin/editor.comparison-grid-row.description',
+  type: 'object',
+  properties: {
+    productFieldsToHide: {
+      title:
+        'admin/editor.comparison-grid-row.product-fields-to-be-removed.title',
+      description:
+        'admin/editor.comparison-grid-row.product-fields-to-be-removed.description',
+      type: 'string',
+      widget: {
+        'ui:widget': 'textarea',
+      },
+    },
+    skuFieldsToHide: {
+      title: 'admin/editor.comparison-grid-row.sku-fields-to-be-removed.title',
+      description:
+        'admin/editor.comparison-grid-row.sku-fields-to-be-removed.description',
+      type: 'string',
+      widget: {
+        'ui:widget': 'textarea',
+      },
+    },
+    productSpecificationsToHide: {
+      title:
+        'admin/editor.comparison-grid-row.product-specifications-to-be-removed.title',
+      description:
+        'admin/editor.comparison-grid-row.product-specifications-to-be-removed.description',
+      type: 'string',
+      widget: {
+        'ui:widget': 'textarea',
+      },
+    },
+    skuSpecificationsToHide: {
+      title:
+        'admin/editor.comparison-grid-row.sku-specifications-to-be-removed.title',
+      description:
+        'admin/editor.comparison-grid-row.sku-specifications-to-be-removed.description',
+      type: 'string',
+      widget: {
+        'ui:widget': 'textarea',
+      },
+    },
+  },
 }
 
 export default ComparisonGrid