+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Product comparison hide fields
+
 ### Added
 - Product comparison grid basic feature
 - Comparison list normalize product for product summary